Refactor GNSS_IMU_Fusion into modular package
